refactor: posts관련 api 분리
